udpate framework

+++ Test/app/src/main/jni/mediasdk/include/engine.h
@@ -49,7 +49,7 @@ namespace paomiantv {
 
         void setDataSource(CStoryboard *pStoryboard);
 
-        void seekTo(s64 sllPosition);
+        void seekTo(s32 sllPosition);
